修复Dockerfile的问题，解除脚本中的错误及警告

+++ scripts/deploy.sh
@@ -1,9 +1,10 @@
+#!/bin/sh
 # update api blueprint documet
 if [ -d /opt/api-blueprint ]
 then
-    cd /opt/api-blueprint
+    cd /opt/api-blueprint || exit
 
-    if [[ -n $repository ]]; then
+    if [[ -n "$(repository)" ]]; then
         echo "git pull $repository"
         git checkout -f
         git clean -f
@@ -13,7 +14,7 @@ then
 else
         echo "git clone $repository"
         git clone $repository /opt/api-blueprint
-        cd /opt/api-blueprint
+        cd /opt/api-blueprint || return
 fi
 
 # convert tab and space
@@ -21,16 +22,15 @@ sed -i 's/\t/\ \ /g' *.apib
 chmod 644 *.apib
 
 # build api document
-find . -name "*.apib" | sed 's/.apib//'|xargs -i -t aglio -i {}.apib `echo $aglio` -o {}.html
+find . -name "*.apib" | sed 's/.apib//' |xargs -i -t aglio -i {}.apib `echo "$(aglio)"` -o {}.html
 
 # cp source api docs 2 nginx serve directroy
 dst=/usr/share/nginx/html
-rm -rf $dst/*
-for i in `find . | grep -v '\.git' | grep '\.html$'`; do
+rm -rf "${dst:?}/"*
+for i in `find . -name "*.html"`; do
         mkdir -p $dst/`dirname $i`;
         cp -f $i $dst/$i;
 done
-
 
 # restart drakov
 pkill -9 drakov
